feat: add Prisma ORM and add table atletica
Since the backend routes need a database to store their data, it was necessary to make it a simple way to interact with these tables without having it installed on your computer.

This commit adds a new database container via Docker and Prisma connecting the database with the backend endpoints. It also adds the base migration where the 'atletica' table was added. Besides that, it creates a prisma lib to connect with the backend and it also adds a script that runs the Docker container with some options.